# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Berita Documentation",
        "description": "Dokumentasi API untuk sistem manajemen berita",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/users/{id_user}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get single user",
                "description": "Mengambil detail user berdasarkan ID",
                "operationId": "163dadaa4e56547be21e2297b0992bf6",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "description": "ID user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_user": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "johndoe"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "pembaca"
                                        },
                                        "membership": {
                                            "type": "string",
                                            "example": "free"
                                        },
                                        "foto_profil": {
                                            "type": "string",
                                            "example": "profile.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user profile",
                "description": "Mengupdate username, name, email, password, dan foto profil",
                "operationId": "d74471d6414b21c44f6aa5e5497cfbbc",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "description": "ID user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "old_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "P@ssW0rd3"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newP@ssW0rd3"
                                    },
                                    "foto_profil": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_user": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "johndoe"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "john@example.com"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "pembaca"
                                        },
                                        "membership": {
                                            "type": "string",
                                            "example": "free"
                                        },
                                        "foto_profil": {
                                            "type": "string",
                                            "example": "profile.jpg"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Password lama salah"
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user",
                "description": "Menghapus user",
                "operationId": "26b1d3f0d3f3d223b97c69f73802c588",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "description": "ID user",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User berhasil dihapus"
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter your Bearer token",
                "bearerFormat": "JWT"
            }
        }
    },
...
}